chore: refactored gpt assistant modal to popover component
