+++ web/components/inbox/modals/create-issue-modal.tsx
@@ -16,7 +16,7 @@ import { Button, Input, ToggleSwitch } from "@plane/ui";
 // types
 import { IIssue } from "types";
 import useEditorSuggestions from "hooks/use-editor-suggestions";
-import { GptAssistantModal } from "components/core";
+import { GptAssistantPopover } from "components/core";
 import { Sparkle } from "lucide-react";
 import useToast from "hooks/use-toast";
 import { AIService } from "services/ai.service";
@@ -227,7 +227,7 @@ export const CreateInboxIssueModal: React.FC<Props> = observer((props) => {
                           />
                         </div>
                         <div className="relative">
-                          <div className="flex justify-end">
+                          <div className="border-0.5 absolute bottom-3.5 right-3.5 z-10 flex rounded bg-custom-background-80">
                             {issueName && issueName !== "" && (
                               <button
                                 type="button"
@@ -246,14 +246,33 @@ export const CreateInboxIssueModal: React.FC<Props> = observer((props) => {
                                 )}
                               </button>
                             )}
-                            <button
-                              type="button"
-                              className="flex items-center gap-1 rounded px-1.5 py-1 text-xs hover:bg-custom-background-90"
-                              onClick={() => setGptAssistantModal((prevData) => !prevData)}
-                            >
-                              <Sparkle className="h-4 w-4" />
-                              AI
-                            </button>
+
+                            {envConfig?.has_openai_configured && (
+                              <GptAssistantPopover
+                                isOpen={gptAssistantModal}
+                                projectId={projectId}
+                                handleClose={() => {
+                                  setGptAssistantModal((prevData) => !prevData);
+                                  // this is done so that the title do not reset after gpt popover closed
+                                  reset(getValues());
+                                }}
+                                onResponse={(response) => {
+                                  handleAiAssistance(response);
+                                }}
+                                button={
+                                  <button
+                                    type="button"
+                                    className="flex items-center gap-1 rounded px-1.5 py-1 text-xs hover:bg-custom-background-90"
+                                    onClick={() => setGptAssistantModal((prevData) => !prevData)}
+                                  >
+                                    <Sparkle className="h-4 w-4" />
+                                    AI
+                                  </button>
+                                }
+                                className="!min-w-[38rem]"
+                                placement="top-end"
+                              />
+                            )}
                           </div>
                           <Controller
                             name="description_html"
@@ -276,26 +295,6 @@ export const CreateInboxIssueModal: React.FC<Props> = observer((props) => {
                               />
                             )}
                           />
-                          {envConfig?.has_openai_configured && (
-                            <div
-                              className={`absolute top-2 left-0 z-20 flex w-full flex-col space-y-4 overflow-hidden rounded-[10px] border border-custom-border-200 bg-custom-background-100 p-4 shadow ${
-                                gptAssistantModal ? "block" : "hidden"
-                              }`}
-                            >
-                              <GptAssistantModal
-                                isOpen={gptAssistantModal}
-                                handleClose={() => {
-                                  setGptAssistantModal(false);
-                                  // this is done so that the title do not reset after gpt popover closed
-                                  reset(getValues());
-                                }}
-                                onResponse={(response) => {
-                                  handleAiAssistance(response);
-                                }}
-                                projectId={projectId}
-                              />
-                            </div>
-                          )}
                         </div>
 
                         <div className="flex flex-wrap items-center gap-2">
