+++ web/components/issues/draft-issue-form.tsx
@@ -8,7 +8,7 @@ import { FileService } from "services/file.service";
 import useToast from "hooks/use-toast";
 import useLocalStorage from "hooks/use-local-storage";
 // components
-import { GptAssistantModal } from "components/core";
+import { GptAssistantPopover } from "components/core";
 import { ParentIssuesListModal } from "components/issues";
 import {
   IssueAssigneeSelect,
@@ -389,7 +389,7 @@ export const DraftIssueForm: FC<IssueFormProps> = observer((props) => {
               )}
               {(fieldsToShow.includes("all") || fieldsToShow.includes("description")) && (
                 <div className="relative">
-                  <div className="flex justify-end">
+                  <div className="border-0.5 absolute bottom-3.5 right-3.5 z-10 flex rounded bg-custom-background-80">
                     {issueName && issueName !== "" && (
                       <button
                         type="button"
@@ -408,14 +408,32 @@ export const DraftIssueForm: FC<IssueFormProps> = observer((props) => {
                         )}
                       </button>
                     )}
-                    <button
-                      type="button"
-                      className="flex items-center gap-1 rounded px-1.5 py-1 text-xs hover:bg-custom-background-90"
-                      onClick={() => setGptAssistantModal((prevData) => !prevData)}
-                    >
-                      <Sparkle className="h-4 w-4" />
-                      AI
-                    </button>
+                    {envConfig?.has_openai_configured && (
+                      <GptAssistantPopover
+                        isOpen={gptAssistantModal}
+                        projectId={projectId}
+                        handleClose={() => {
+                          setGptAssistantModal((prevData) => !prevData);
+                          // this is done so that the title do not reset after gpt popover closed
+                          reset(getValues());
+                        }}
+                        onResponse={(response) => {
+                          handleAiAssistance(response);
+                        }}
+                        button={
+                          <button
+                            type="button"
+                            className="flex items-center gap-1 rounded px-1.5 py-1 text-xs hover:bg-custom-background-90"
+                            onClick={() => setGptAssistantModal((prevData) => !prevData)}
+                          >
+                            <Sparkle className="h-4 w-4" />
+                            AI
+                          </button>
+                        }
+                        className=" !min-w-[38rem]"
+                        placement="top-end"
+                      />
+                    )}
                   </div>
                   <Controller
                     name="description_html"
@@ -443,26 +461,6 @@ export const DraftIssueForm: FC<IssueFormProps> = observer((props) => {
                       />
                     )}
                   />
-                  {envConfig?.has_openai_configured && (
-                    <div
-                      className={`absolute top-2 left-0 z-20 flex w-full flex-col space-y-4 overflow-hidden rounded-[10px] border border-custom-border-200 bg-custom-background-100 p-4 shadow ${
-                        gptAssistantModal ? "block" : "hidden"
-                      }`}
-                    >
-                      <GptAssistantModal
-                        isOpen={gptAssistantModal}
-                        handleClose={() => {
-                          setGptAssistantModal(false);
-                          // this is done so that the title do not reset after gpt popover closed
-                          reset(getValues());
-                        }}
-                        onResponse={(response) => {
-                          handleAiAssistance(response);
-                        }}
-                        projectId={projectId}
-                      />
-                    </div>
-                  )}
                 </div>
               )}
               <div className="flex flex-wrap items-center gap-2">
