+++ web/components/issues/form.tsx
@@ -10,7 +10,7 @@ import { FileService } from "services/file.service";
 // hooks
 import useToast from "hooks/use-toast";
 // components
-import { GptAssistantModal } from "components/core";
+import { GptAssistantPopover } from "components/core";
 import { ParentIssuesListModal } from "components/issues";
 import {
   IssueAssigneeSelect,
@@ -363,14 +363,31 @@ export const IssueForm: FC<IssueFormProps> = observer((props) => {
                         )}
                       </button>
                     )}
-                    <button
-                      type="button"
-                      className="flex items-center gap-1 rounded px-1.5 py-1 text-xs hover:bg-custom-background-90"
-                      onClick={() => setGptAssistantModal((prevData) => !prevData)}
-                    >
-                      <Sparkle className="h-4 w-4" />
-                      AI
-                    </button>
+                    {envConfig?.has_openai_configured && (
+                      <GptAssistantPopover
+                        isOpen={gptAssistantModal}
+                        projectId={projectId}
+                        handleClose={() => {
+                          setGptAssistantModal((prevData) => !prevData);
+                          // this is done so that the title do not reset after gpt popover closed
+                          reset(getValues());
+                        }}
+                        onResponse={(response) => {
+                          handleAiAssistance(response);
+                        }}
+                        placement="top-end"
+                        button={
+                          <button
+                            type="button"
+                            className="flex items-center gap-1 rounded px-1.5 py-1 text-xs hover:bg-custom-background-90"
+                            onClick={() => setGptAssistantModal((prevData) => !prevData)}
+                          >
+                            <Sparkle className="h-4 w-4" />
+                            AI
+                          </button>
+                        }
+                      />
+                    )}
                   </div>
                   <Controller
                     name="description_html"
@@ -398,26 +415,6 @@ export const IssueForm: FC<IssueFormProps> = observer((props) => {
                       />
                     )}
                   />
-                  {envConfig?.has_openai_configured && (
-                    <div
-                      className={`absolute top-2 left-0 z-20 flex w-full flex-col space-y-4 overflow-hidden rounded-[10px] border border-custom-border-200 bg-custom-background-100 p-4 shadow ${
-                        gptAssistantModal ? "block" : "hidden"
-                      }`}
-                    >
-                      <GptAssistantModal
-                        isOpen={gptAssistantModal}
-                        handleClose={() => {
-                          setGptAssistantModal(false);
-                          // this is done so that the title do not reset after gpt popover closed
-                          reset(getValues());
-                        }}
-                        onResponse={(response) => {
-                          handleAiAssistance(response);
-                        }}
-                        projectId={projectId}
-                      />
-                    </div>
-                  )}
                 </div>
               )}
               <div className="flex flex-wrap items-center gap-2">
